engine: extract dispatch::act_agenda (#159)
